% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeActivScore.R
\name{activeScorePCAlist}
\alias{activeScorePCAlist}
\title{Compute activation score for a list of gene sets.}
\usage{
activeScorePCAlist(
  exprMatrix,
  geneList,
  transpose = TRUE,
  scale = FALSE,
  center = TRUE
)
}
\arguments{
\item{exprMatrix}{A matrix of expression values, genes as rows and samples as columns.}

\item{geneList}{A list of gene sets, each element is a vector of genes that exist in \code{exprMatrix} row names.}

\item{transpose}{Logical. If TRUE, \code{exprMatrix} is transposed with \code{t()} before computing the PCA.}

\item{scale}{Logical. Divide expression of gene by its standard deviation before doing the PCA.}

\item{center}{Logical. Subtract mean to gene expression before doing the PCA.}
}
\value{
A list with two elements:
\itemize{
\item activScoreMat: A matrix of activation score, with gene sets as rows and samples as columns.
\item contributionList: A list of contribution (or weight) to activation score of each gene per gene set.
}
}
\description{
Compute activation score for a list of gene sets.
}
\examples{
data("bulkLogCounts", package = "oob")
keggDB<-getDBterms(rownames(bulkLogCounts),database = "kegg")
geneSetActivScore<-activeScorePCAlist(bulkLogCounts,geneList = keggDB[[1]])
}
