% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeActivScore.R
\name{activScorePC1}
\alias{activScorePC1}
\title{Compute the activation score of a gene set from 1st component of its PCA}
\usage{
activScorePC1(
  data,
  genes,
  transpose = TRUE,
  scale = FALSE,
  center = TRUE,
  returnContribution = FALSE,
  se_data_assay = "logcounts"
)
}
\arguments{
\item{data}{A matrix of numeric with rows as features (in the RNA-Seq
context, log counts). Can also be a SummarizedExperiment object.}

\item{genes}{A character vector. The gene set where the activation score has
to be computed. Must be a subset of \code{data} row names.}

\item{transpose}{Logical. If TRUE, \code{data} is transposed with \code{t()} before
computing the PCA.}

\item{scale}{Logical. Divide features by their standard deviation.}

\item{center}{Logical. Subtract features by their average.}

\item{returnContribution}{Logical. Return list with activation score and
contribution of genes to the activation score.}

\item{se_data_assay}{Integer
or character, if \code{data} is a \code{SummarizedExperiment} object, the assay name
to use.}
}
\value{
A vector of numeric corresponding to activation scores, named by genes. If \code{returnContribution} return a list with activation scores and contributions of genes.
}
\description{
Compute the activation score of a gene set from 1st component of its PCA
}
\examples{
data("bulkLogCounts")
data("keggHuman")
geneSet<-keggHuman$kegg$`hsa00190 Oxidative phosphorylation`
geneSet<-intersect(geneSet,rownames(bulkLogCounts))
activScorePC1(bulkLogCounts,genes = geneSet)
activScorePC1(bulkLogCounts,genes = geneSet,returnContribution = TRUE)
}
