% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeActivScore.R
\name{activScorePC1list}
\alias{activScorePC1list}
\title{Compute activation score for a list of gene sets.}
\usage{
activScorePC1list(
  data,
  geneList,
  transpose = TRUE,
  scale = FALSE,
  center = TRUE,
  BPPARAM = BiocParallel::SerialParam()
)
}
\arguments{
\item{data}{A matrix of expression values, genes as rows and samples as columns. Can also be a SummarizedExperiment object.}

\item{geneList}{A list of gene sets, each element is a vector of genes that exist in \code{data} row names.}

\item{transpose}{Logical. If TRUE, \code{data} is transposed with \code{t()} before computing the PCA.}

\item{scale}{Logical. Divide expression of gene by its standard deviation before doing the PCA.}

\item{center}{Logical. Subtract mean to gene expression before doing the PCA.}

\item{BPPARAM}{Parallelization parameter as used in BiocParallel::bpparam().}
}
\value{
A list with two elements:
\itemize{
\item activScoreMat: A matrix of activation score, with gene sets as rows and samples as columns.
\item contributionList: A list of contribution (or weight) to activation score of each gene per gene set.
}
}
\description{
Compute activation score for a list of gene sets.
}
\examples{
data("bulkLogCounts")
data("keggHuman")
geneList <-  keggHuman$kegg[seq.int(5)]
geneList <- lapply(geneList, function(x) intersect(x,rownames(bulkLogCounts)))
geneSetActivScore<-activScorePC1list(bulkLogCounts,geneList = geneList)
}
