% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeActivScore.R
\name{computeActivationScore}
\alias{computeActivationScore}
\title{Compute the activation score of gene sets from an expression matrix.}
\usage{
computeActivationScore(
  data,
  DBsets,
  transpose = TRUE,
  scaleScores = FALSE,
  returnSummarizedExperiment = TRUE,
  maxSize = 500,
  minSize = 2,
  se_data_assay = "logcounts",
  BPPARAM = BiocParallel::SerialParam()
)
}
\arguments{
\item{data}{An expression matrix (normalized log2(x+1) counts). Genes as rows
and sample as columns by default. If \code{db_terms} is not given, must be named
by gene symbols. Can also be a SummarizedExperiment object.}

\item{DBsets}{A list or NULL. A named list were each element is a database.
Inside each database, a list terms, named by the term and containing gene
vectors as gene symbols.}

\item{transpose}{Logical. If \code{TRUE}, \code{data} is transposed with \code{t()} before
computing the PCA.}

\item{scaleScores}{Logical. Divide expression of gene by its standard
deviation before doing the PCA.}

\item{returnSummarizedExperiment}{do not keep the results listed by database but merge them as a single activation SummarizedExperiment object}

\item{maxSize}{Maximum number of gene in each term.}

\item{minSize}{Minimum number of gene in each term.}

\item{se_data_assay}{Integer
or character, if \code{data} inherit a \code{SummarizedExperiment} object, the assay name where the input matrix is.}

\item{BPPARAM}{Parallelization parameter as used in BiocParallel::bpparam().}
}
\value{
if \code{returnSummarizedExperiment = TRUE}
A SummarizedExperiment object, with the activation score as the assay, and gene set metadata stored in the \code{rowData}.
The gene set metadata contains size_db (number of gene in the term originally present in the given database),
size_universe (number of gene in the term after removing genes not present).
if \code{returnSummarizedExperiment = FALSE}
A list by database containing each a list with the matrix of activation score, with gene sets as
columns and samples as rows ; and the list of contribution (or weight) to
activation score of each gene per gene set, these two elements are superseded by the list of database given as input.
}
\description{
Perform a PCA for each gene set, from the matrix of \emph{genes from
gene set × all samples}. Return the first PCs as activation scores of the
gene sets.
}
\examples{
data("bulkLogCounts")
data("keggHuman")
geneSetActivScore<-computeActivationScore(bulkLogCounts, keggHuman)
}
