% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manageDB.R
\name{downloadGODB_BioMart}
\alias{downloadGODB_BioMart}
\title{Download GO term for a particular Ensembl dataset, useful for non model species}
\usage{
downloadGODB_BioMart(
  ensemblDataset = "mzebra_gene_ensembl",
  returnGeneSymbol = FALSE
)
}
\arguments{
\item{ensemblDataset}{name (species) of the Ensembl dataset}

\item{returnGeneSymbol}{Return gene symbol if \code{TRUE} or Ensembl IDs}
}
\value{
A list where each element is a database that contain a list of term with the associated gene symbols.
}
\description{
Download GO term for a particular Ensembl dataset, useful for non model species
}
\examples{
\dontrun{
dbGO_Mzebra <- downloadGODB_BioMart("mzebra_gene_ensembl")
}
}
