% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeEnrichment.R
\name{enrich.fcs}
\alias{enrich.fcs}
\title{Functional class scoring enrichment (fgsea algorithm)}
\usage{
enrich.fcs(
  x,
  DBsets,
  maxSize = 500,
  minSize = 2,
  returnGenes = FALSE,
  BPPARAM = BiocParallel::SerialParam(),
  ...
)
}
\arguments{
\item{x}{vector or dataframe/matrix of one column. Values are used to classify the genes, example, it can be Log2(Fold-Change0). Genes are contained in the names/rownames of the vector/dataframe. Example of valid x: x<-rnorm(n = 4); names(x)<-c("GATA2","SOX17","KLF4","POU5F1")}

\item{DBsets}{A list or NULL. A named list were each element is a database.
Inside each database, a list terms, named by the term and containing gene
vectors as gene symbols.}

\item{maxSize}{Maximum number of gene in each term.}

\item{minSize}{Minimum number of gene in each term.}

\item{returnGenes}{return genes that were the most important for the enrichment of term}

\item{BPPARAM}{Parallelization parameter as used in BiocParallel::bpparam().}

\item{...}{Additional parameters that are passed to fgsea}
}
\value{
A data frame with the following columns:
\itemize{
\item term: name of the term/gene set
\item database: origin of the gene set
\item size: number of gene in the term after removing genes not present.
\item pval: an enrichment p-value
\item padj: a BH-adjusted p-value
\item log2err: the expected error for the standard deviation of the P-value logarithm
\item ES: enrichment score, same as in Broad GSEA implementation
\item NES: enrichment score normalized to mean enrichment of random samples of the same size
\item genes (if \code{returnGenes}). Vector of genes of the term.
}
}
\description{
Functional class scoring enrichment (fgsea algorithm)
}
\examples{
data("DEgenesPrime_Naive")
data("keggHuman")
fcsScore<-fcsScoreDEgenes(rownames(DEgenesPrime_Naive),
    DEgenesPrime_Naive$pvalue,DEgenesPrime_Naive$log2FoldChange)
resEnrich<-enrich.fcs(fcsScore,DBsets =  keggHuman)
head(resEnrich)
}
