% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manageDB.R
\name{exportEnrich}
\alias{exportEnrich}
\title{Export enrichment results with the gene list associated to each term/row.}
\usage{
exportEnrich(
  enrichResults,
  file,
  sep = "\\t",
  col.names = TRUE,
  row.names = FALSE,
  geneCol = "genes",
  ...
)
}
\arguments{
\item{enrichResults}{Dataframe, usually from Enrich functions (for example \code{enrich.ora}) with \code{returnGenes=TRUE}.}

\item{file}{Path to file to write.}

\item{sep}{Field separator.}

\item{col.names}{Either a logical value indicating whether the column names of x are to be written along with x, or a character vector of column names to be written. See the section on ‘CSV files’ for the meaning of col.names = NA.}

\item{row.names}{Either a logical value indicating whether the row names of x are to be written along with x, or a character vector of row names to be written.}

\item{geneCol}{The column that contain the list of gene of the term.}

\item{...}{Other parameters passed to write.table.}
}
\value{
Write a text file.
}
\description{
Export enrichment results with the gene list associated to each term/row.
}
