% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manageDB.R
\name{getDBsets}
\alias{getDBsets}
\title{Download database of term/pathway for enrichment analyses.}
\usage{
getDBsets(
  species = "Human",
  database = c("kegg", "reactom", "goBP", "goCC", "goMF"),
  id_type = "SYMBOL",
  GOALL = FALSE,
  speciesData = NULL
)
}
\arguments{
\item{species}{Character. Shortname of the species as described in \code{data("species_df")}.}

\item{database}{Which annotation database ? valid: database: kegg reactom goBP goCC goMF custom.}

\item{id_type}{Character. Gene ID type that has to be returned, typically values can be "ENSEMBL", "ENTREZID" or "SYMBOL", see columns(speciesData$getSpeciesPackage()) to see available IDs}

\item{GOALL}{Logical. If \code{TRUE} the GO term correspond to propagated annotations for all parents term, which make the database huge an less specific, but more complete.}

\item{speciesData}{object returned by \code{getSpeciesData}. If not NULL \code{species} argument wont be used.}
}
\value{
A list where each element is a database that contain a list of term with the associated gene symbols.
}
\description{
Download database of term/pathway for enrichment analyses.
}
\examples{
data("bulkLogCounts")
enrichDBs<-getDBsets(species="Human",database=c("kegg","goBP"))
}
