% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeEnrichment.R
\name{getMarkers}
\alias{getMarkers}
\title{Compute a dataframe with marker metrics describing best gene marker per group
of samples.}
\usage{
getMarkers(
  data,
  groups,
  transpose = TRUE,
  BPPARAM = BiocParallel::SerialParam(),
  returnAsList = FALSE,
  se_assay = "logcounts"
)
}
\arguments{
\item{data}{A dataframe with genes as rows and samples as columns.
Can also be a \code{SummarizedExperiment} or \code{SummarizedExperiment} object.}

\item{groups}{A vector of group names, same size as the number of columns in
\code{data}.}

\item{transpose}{If TRUE, the input data is transposed before processing.
Default is TRUE (feature as rows, samples as columns).}

\item{BPPARAM}{A BPPARAM object as return by \code{\link[BiocParallel:register]{BiocParallel::bpparam()}}. Used
for multi-threading.}

\item{returnAsList}{Return a list where each element is dataframe containing
the marker metrics of a group.}

\item{se_assay}{Integer or character, if \code{data} is a
\code{SummarizedExperiment} related object, the assay name to use.}
}
\value{
A dataframe containing four column per group: Log2(Fold-Change),
AUROC, marker score (see details), p-value and BH adjusted p-value.
If \code{data} is a \code{SummarizedExperiment} related object and \code{returnAsList} is
\code{FALSE}, the function will add the marker metrics to \code{rowData}.
}
\description{
Compute a dataframe with marker metrics describing best gene marker per group
of samples.
}
\details{
LogFC and pvalues are computed from a linear modelling of the data.

Score is consisting of the geometrical mean of absolute LogFC, absolute(auroc
\itemize{
\item 0.5), and -log10(pval), then signed by the logFC: score = sign(logFC) ×
gmean( abs(logFC), abs(aurocRes-0.5), -log10(pval) )
}
}
\examples{
data("bulkLogCounts")
data("sampleAnnot")
markerData <- getMarkers(bulkLogCounts,sampleAnnot$culture_media)
sce <- SummarizedExperiment::SummarizedExperiment(assays = list(counts = bulkLogCounts))
sce <- getMarkers(sce, sampleAnnot$culture_media, se_assay = "counts")
rowData(sce) |> head()
}
