% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manageDB.R
\name{getSpeciesData}
\alias{getSpeciesData}
\title{Return gene id correspondence, GO species code and KEGG species code}
\usage{
getSpeciesData(species = "Human", species_df = NULL)
}
\arguments{
\item{species}{Character. Shortname of the species (rownames in \code{data("species_df")}).}

\item{species_df}{Eventually an updated \code{species_df} dataframe.}
}
\value{
A list describing specific data for the species (gene IDs, annotation package...).
}
\description{
Return gene id correspondence, GO species code and KEGG species code
}
\examples{
require(BiocManager)
require(org.Hs.eg.db)

getSpeciesData("Human") |> str()

#Available species
data("species_df")
species_df[,"species"]
}
