% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.R
\name{GSDS.Heatmap}
\alias{GSDS.Heatmap}
\title{Plot the activation scores and the top gene contributions in one heatmap}
\usage{
GSDS.Heatmap(
  GSDS.object,
  colData = NULL,
  row_names_max_width = unit(340, "points"),
  maxgene_shownperside = 3,
  contrib_width = NULL,
  contrib_max_width = unit(500, "points"),
  contrib_fontface = "italic",
  contrib_fontsize = NULL,
  contrib_additional_gp = list(),
  ...
)
}
\arguments{
\item{GSDS.object}{A \code{SummarizedExperiment} as returned by \code{computeActivationScore} or \code{GSDS}}

\item{colData}{A vector of factor, character, numeric or logical. Or, a
dataframe of any of these type of value. The annotation that will be
displayed on the heatmap.}

\item{row_names_max_width}{Unit. Max width of the heatmap gene set}

\item{maxgene_shownperside}{Integer. Number of best pos/neg rank displayed on the annotation}

\item{contrib_width}{Unit. Width of the gene contribution annotation. If NULL computed automatically}

\item{contrib_max_width}{Unit. If width is computed automatically, maximum width of the contribution annotation}

\item{contrib_fontface}{Character. Fontface of the genes contribution annotation}

\item{contrib_fontsize}{Unit. Fontsize of the genes contribution annotation. If NULL computed automatically}

\item{contrib_additional_gp}{Additional gpar  argument passed as a list to the contribution annotation, see ?grid::gpar}

\item{...}{Arguments passed to HEATmap}
}
\value{
A heatmap of the activation scores, names by gene sets, and with the top gene contributions on the right.
a selection of genes with a positive or negative contribution to the activation score is represented each set.
Positive contribution (at the right of the vertical bar), means that up regulation increases the activation score, whereas
a negative contribution (left of the bar) means that an increase of expression will decrease the activation score.
This makes it easy to interpret whether a positive activation score means that a pathway is "biologically" on or off,
or assess the redundance of gene sets.
}
\description{
Plot the activation scores and the top gene contributions in one heatmap
}
\examples{
data("bulkLogCounts")
data("sampleAnnot")
data("keggHuman")

resGSDS <- GSDS(
	data = bulkLogCounts,
	DBsets = keggHuman,
	colData = sampleAnnot,
	contrast = c("culture_media", "T2iLGO", "KSR+FGF2"),
	return_seobject = TRUE
)
resGSDS.sel <- resGSDS[abs(rowData(resGSDS)$log2FoldChange) > 7]

GSDS.Heatmap(resGSDS.sel,  colData = sampleAnnot["culture_media"])

}
\seealso{
\code{\link[=HEATmap]{HEATmap()}}, \code{\link[=GSDS.HeatmapAnnot]{GSDS.HeatmapAnnot()}}
}
