% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.R
\name{GSDS.HeatmapAnnot}
\alias{GSDS.HeatmapAnnot}
\title{Displaying most neg and pos gene contribution on a GSDS heatmap.}
\usage{
GSDS.HeatmapAnnot(
  contributions,
  maxgene_shownperside = 3,
  width = NULL,
  max_width = unit(500, "points"),
  fontface = "italic",
  fontsize = NULL,
  additional_gp = list()
)
}
\arguments{
\item{contributions}{A list named by gene set Contains vector of gene contribution to activation scores, named by genes.}

\item{maxgene_shownperside}{Integer. Number of best pos/neg rank displayed on the annotation}

\item{width}{Unit. Width of the annotation heatmap. If NULL computed automatically}

\item{max_width}{Unit. If width is computed automatically, maximum width of the contribution annotation}

\item{fontface}{Character. Fontface of the genes text}

\item{fontsize}{Unit. Fontsize of the genes text. If NULL computed automatically}

\item{additional_gp}{Additional gpar argument passed as a list, see ?grid::gpar}
}
\value{
A HeatmapAnnotation object. Genes on the left side of the vertical line are contributing negatively to the activation score, and positively on the right side.
}
\description{
Displaying most neg and pos gene contribution on a GSDS heatmap.
}
\examples{
data("bulkLogCounts")
data("sampleAnnot")
data("keggHuman")

resGSDS <- GSDS(
	data = bulkLogCounts,
	DBsets = keggHuman,
	colData = sampleAnnot,
	contrast = c("culture_media", "T2iLGO", "KSR+FGF2"),
	return_seobject = TRUE
)
resGSDS.sel <- resGSDS[abs(rowData(resGSDS)$log2FoldChange) > 10]

nrow(resGSDS.sel)

HEATmap(
		assay(resGSDS.sel),
    midColorIs0= TRUE,
    center = FALSE,
    name = "Activation score",
    preset = NULL,
    colData = sampleAnnot["culture_media"],
    right_annotation = ComplexHeatmap::rowAnnotation(
        "gene contribution" =
            GSDS.HeatmapAnnot(
                contributions = rowData(resGSDS)$contributions,
                fontsize = unit(8,"points")

            )
    ),
    row_names_side = "left",
    row_dend_side = "left",
    row_names_max_width = unit(340, "points")
)
}
