% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeEnrichment.R
\name{GSDS}
\alias{GSDS}
\title{Gene Set Differential Scoring (GSDS)}
\usage{
GSDS(
  geneSetActivScore = NULL,
  DBsets = NULL,
  data = NULL,
  colData,
  contrast,
  formula = NULL,
  maxSize = 500,
  minSize = 2,
  se_data_assay = "logcounts",
  return_seobject = TRUE,
  BPPARAM = BiocParallel::SerialParam()
)
}
\arguments{
\item{geneSetActivScore}{A \code{SummarizedExperiment} as returned by \code{computeActivationScore}}

\item{DBsets}{A list or NULL. A named list were each element is a database.
Inside each database, a list terms, named by the term and containing gene
vectors as gene symbols.}

\item{data}{An expression matrix (normalized log2(x+1) counts). Genes as rows
and sample as columns by default. If \code{db_terms} is not given, must be named
by gene symbols. Can also be a SummarizedExperiment object.}

\item{colData}{An annotation dataframe. Each column is a feature, each row a sample. Same number of samples than in \code{data}.}

\item{contrast}{A vector of 3 character.
\enumerate{
\item Name of the experimental variable that have to be used for differential activation. Must be a column name of \code{colData}.
\item Condition considered as the reference.
\item Condition considered as the target group.
}}

\item{formula}{Formula for the linear modelling, if NULL formula is group. Example : \code{~ Condition + Batch} if condition and
Batch are colnames of colData}

\item{maxSize}{Maximum number of gene in each term.}

\item{minSize}{Minimum number of gene in each term.}

\item{se_data_assay}{Integer
or character, if \code{data} inherit a \code{SummarizedExperiment} object, the assay name where the input matrix is.}

\item{return_seobject}{return a \code{SummarizedExperiment} object, with rowData annotated with GSDS results instead of the result dataframe.
to use.}

\item{BPPARAM}{Parallelization parameter as used in BiocParallel::bpparam().}
}
\value{
A dataframe name by gene set IDs, with the following columns:
\itemize{
\item original_name: name of the term/gene set as it is in the provided database
\item database: origin of the gene set
\item size_db: number of gene in the gene set.
\item size_universe:number of gene in the gene set that are presesent in the data (e.g. used for computing the activation score)
\item baseMean: mean of activation score in the gene set
\item sd: standard deviation  in the gene set
\item log2FoldChange: Log(Log Fold Change) of activation score between the two tested groups.
\item pval: an enrichment p-value
\item padj: a BH-adjusted p-value
}
}
\description{
Gene Set Differential Scoring (GSDS)
}
\examples{
data("bulkLogCounts")
data("sampleAnnot")
data("keggHuman")

geneSetActivScore<-computeActivationScore(bulkLogCounts,DBsets = keggHuman)
resGSDS<-GSDS(geneSetActivScore = geneSetActivScore,colData = sampleAnnot,
    contrast = c("culture_media","T2iLGO","KSR+FGF2"))
head(rowData(resGSDS))
# correct for the line effect
resGSDS<-GSDS(geneSetActivScore = geneSetActivScore,colData = sampleAnnot,
    contrast = c("culture_media","T2iLGO","KSR+FGF2"), formula = ~ culture_media + line)
head(rowData(resGSDS))

#just a data frame
head(GSDS(geneSetActivScore = geneSetActivScore,colData = sampleAnnot,
    contrast = c("culture_media","T2iLGO","KSR+FGF2"),DBsets =  keggHuman,
    return_seobject = FALSE))
# or
resGSDS<-GSDS(data = bulkLogCounts,colData = sampleAnnot,
    contrast = c("culture_media","T2iLGO","KSR+FGF2"),DBsets = keggHuman)

}
