% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{gsds_colors}
\alias{gsds_colors}
\title{Colors for a qualitative scale}
\usage{
gsds_colors(n = 20)
}
\arguments{
\item{n}{Number of wanted colors}
}
\value{
A vector of colors
}
\description{
Colors for a qualitative scale
}
\examples{
gsds_colors() |> scales::show_col()
gsds_colors(n=5) |> scales::show_col()
gsds_colors(n=40) |> scales::show_col()
}
