% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSDS-package.R
\docType{data}
\name{keggHuman}
\alias{keggHuman}
\title{keggDB}
\format{
A list where first element is a list of 369 gene sets
}
\description{
KEGG gene sets for human
}
\references{
\url{https://doi.org/10.1093/nar/28.1.27}
}
\author{
Kanehisa M
}
\keyword{data}
