% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSDS-package.R
\docType{data}
\name{species_df}
\alias{species_df}
\title{species_df}
\format{
A data frame containing 21 rows for 6 columns
}
\description{
A data frame describing available species (available Organism-level Databases),
Columns: package, species, simple_species, kegg_code, id_type, tax_id"
}
\author{
Dimitri Meistermann
}
\keyword{data}
