% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeEnrichment.R
\name{enrich.ora}
\alias{enrich.ora}
\title{Over Representation Analysis from Hypergeometric test}
\usage{
enrich.ora(x, DBsets, minSize = 2, maxSize = 500, returnGenes = FALSE)
}
\arguments{
\item{x}{vector or dataframe/matrix of one column.
Values are booleans and say if gene is from the list of interest or not.
Genes are contained in the names/rownames of the vector/dataframe.
Example of valid x: x<-c(TRUE,TRUE,FALSE,FALSE); names(x)<-c("GATA2","SOX17","KLF4","POU5F1").
In this case, GATA2, SOX17, KLF4, POU5F1 are the universe of gene and GATA2 and SOX17 are the genes of interest}

\item{DBsets}{A list or NULL. A named list were each element is a database.
Inside each database, a list terms, named by the term and containing gene
vectors as gene symbols.}

\item{minSize}{Minimum number of gene in each term.}

\item{maxSize}{Maximum number of gene in each term.}

\item{returnGenes}{return genes that were the most important for the enrichment of term}
}
\value{
A dataframe with the following columns:
\itemize{
\item term: name of the term/gene set
\item database: origin of the gene set
\item size_db: number of gene in the term originally present in the given database.
\item size_universe: number of gene in the term after removing genes not present.
\item obsOverlap: number of gene in common between set of interest and term
\item expOverlap: expected number of gene in common between set of interest and term
\item OEdeviation: \eqn{\frac{obs - exp}{\sqrt{Universe}}}, normalized deviation from expected value
\item log2OR: log2 of the Odds Ratio
\item pval: p-value based on a Hypergeometric test
\item padj: a BH-adjusted p-value
\item genes (if \code{returnGenes}). Vector of genes of the term.
}
}
\description{
Over Representation Analysis from Hypergeometric test
}
\examples{
data("DEgenesPrime_Naive")
data("keggHuman")
vectorIsDE<-DEgenesPrime_Naive$isDE!="NONE";names(vectorIsDE)<-rownames(DEgenesPrime_Naive)
resEnrich<-enrich.ora(vectorIsDE, DBsets = keggHuman)
head(resEnrich)
}
