% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manageDB.R
\name{getSpeciesData}
\alias{getSpeciesData}
\title{Return gene id correspondence, GO species code and KEGG species code}
\usage{
getSpeciesData(sample.species = "Human")
}
\arguments{
\item{sample.species}{Character. Shortname of the species as described in \code{data("bods")}.}
}
\value{
A list describing specific data for the species (gene IDs, annotation package...).
}
\description{
Return gene id correspondence, GO species code and KEGG species code
}
\examples{
require(BiocManager)
require(org.Hs.eg.db)

getSpeciesData("Human") |> str()

#Available species
data("bods", package = "gage")
bods[,"species"]
}
