% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSDS-package.R
\docType{package}
\name{GSDS-package}
\alias{GSDS}
\alias{GSDS-package}
\title{GSDS: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\keyword{internal}
