% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeEnrichment.R
\name{GSDS}
\alias{GSDS}
\title{Gene Set Differential Scoring (GSDS)}
\usage{
GSDS(
  geneSetActivScore = NULL,
  exprMatrix = NULL,
  colData,
  contrast,
  idGeneDF = NULL,
  database = c("kegg", "reactom", "goBP", "goCC", "goMF"),
  maxSize = 500,
  minSize = 2,
  customAnnot = NULL,
  keggDisease = FALSE,
  species = "Human",
  db_terms = NULL,
  speciesData = NULL
)
}
\arguments{
\item{geneSetActivScore}{A list of database with an element "eigen" containing the matrix of gene set activation score (see what \code{computeActivationScore} returns).
If NULL, this is computed automatically from the \code{exprMatrix} and the gene set database given via \code{db_terms} or requested via \code{database}.}

\item{exprMatrix}{An expression matrix (normalized log2(x+1) counts). Genes as rows and sample as columns. If \code{db_terms} is not given, must be named by gene symbols.}

\item{colData}{An annotation dataframe. Each column is a feature, each row a sample. Same number of samples than in \code{exprMatrix}.}

\item{contrast}{A vector of 3 character.
\enumerate{
\item Name of the experimental variable that have to be used for differential activation. Must be a column name of \code{colData}.
\item Condition considered as the reference.
\item Condition considered as the target group.
}}

\item{idGeneDF}{Dataframe of gene ID correspondence where each column is a gene ID type. If not NULL \code{species} and \code{speciesData} arguments wont be used.}

\item{database}{Which annotation database ? valid: database: kegg reactom goBP goCC goMF custom.}

\item{maxSize}{Maximum number of gene in each term.}

\item{minSize}{Minimum number of gene in each term.}

\item{customAnnot}{Custom annotation database, as a list of terms, each element contain a vector of gene symbols.}

\item{keggDisease}{Logical. Retain kegg disease term in kegg database?}

\item{species}{Character. Shortname of the species as described in \code{data("bods")}.}

\item{db_terms}{A list or NULL. A named list were each element is a database. Inside each database, a list terms, named by the term and containing gene vectors as gene symbols.}

\item{speciesData}{object returned by \code{getSpeciesData}. If not NULL \code{species} argument wont be used.}
}
\value{
A dataframe with the following columns:
\itemize{
\item term: name of the term/gene set
\item baseMean: mean of activation score in the gene set
\item sd: standard deviation  in the gene set
\item log2FoldChange: Log(Log Fold Change) of activation score between the two tested groups.
\item pval: an enrichment p-value
\item padj: a BH-adjusted p-value
\item database: origin of the gene set
\item size: number of gene in the term after removing genes not present.
}
}
\description{
Gene Set Differential Scoring (GSDS)
}
\examples{
data("bulkLogCounts", package = "oob")
data("sampleAnnot", package = "oob")

keggDB<-getDBterms(rownames(bulkLogCounts),database = "kegg")
geneSetActivScore<-computeActivationScore(bulkLogCounts,db_terms = keggDB)
resGSDS<-GSDS(geneSetActivScore = geneSetActivScore,colData = sampleAnnot,
    contrast = c("culture_media","T2iLGO","KSR+FGF2"),db_terms =  keggDB)

# or
resGSDS<-GSDS(exprMatrix = bulkLogCounts,colData = sampleAnnot,
    contrast = c("culture_media","T2iLGO","KSR+FGF2"),database = "kegg")

head(resGSDS)
}
